refactor: Move scan code into its own package
