+++ pkg/osvscanner/internal/scanners/fixtures/example-git/a.txt
