+++ pkg/osvscanner/internal/scanners/fixtures/example-git/git-hidden/COMMIT_EDITMSG
