+++ pkg/osvscanner/internal/scanners/fixtures/example-git/git-hidden/description
