+++ pkg/osvscanner/internal/scanners/fixtures/example-git/git-hidden/logs/refs/heads/main
