+++ pkg/osvscanner/internal/scanners/fixtures/example-git/git-hidden/objects/4b/825dc642cb6eb9a060e54bf8d69288fbee4904
