+++ pkg/osvscanner/internal/scanners/fixtures/example-git/git-hidden/refs/heads/main
