+++ pkg/osvscanner/vulnerability_result.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/google/osv-scanner/internal/config"
 	"github.com/google/osv-scanner/internal/grouper"
+	"github.com/google/osv-scanner/internal/imodels"
 	"github.com/google/osv-scanner/internal/output"
 	"github.com/google/osv-scanner/internal/sourceanalysis"
 	"github.com/google/osv-scanner/internal/spdx"
@@ -21,7 +22,7 @@ import (
 // TODO: This function is getting long, we should refactor it
 func buildVulnerabilityResults(
 	r reporter.Reporter,
-	packages []scannedPackage,
+	packages []imodels.ScannedPackage,
 	vulnsResp *osv.HydratedBatchedResponse,
 	licensesResp [][]models.License,
 	actions ScannerActions,
