+++ pkg/osvscanner/vulnerability_result_internal_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/google/osv-scanner/internal/config"
+	"github.com/google/osv-scanner/internal/imodels"
 	"github.com/google/osv-scanner/internal/testutility"
 	"github.com/google/osv-scanner/pkg/lockfile"
 	"github.com/google/osv-scanner/pkg/models"
@@ -15,13 +16,13 @@ func Test_assembleResult(t *testing.T) {
 	t.Parallel()
 	type args struct {
 		r            reporter.Reporter
-		packages     []scannedPackage
+		packages     []imodels.ScannedPackage
 		vulnsResp    *osv.HydratedBatchedResponse
 		licensesResp [][]models.License
 		actions      ScannerActions
 		config       *config.Manager
 	}
-	packages := []scannedPackage{
+	packages := []imodels.ScannedPackage{
 		{
 			Name:      "pkg-1",
 			Ecosystem: lockfile.Ecosystem("npm"),
